Move the auth channel to a separate file for imports
