+++ src/store/authentication/channels.ts
@@ -1,0 +1,15 @@
+import { multicastChannel } from 'redux-saga';
+import { call } from 'redux-saga/effects';
+
+export enum Events {
+  UserLogin = 'user/login',
+  UserLogout = 'user/logout',
+}
+
+let theChannel;
+export function* authChannel() {
+  if (!theChannel) {
+    theChannel = yield call(multicastChannel);
+  }
+  return theChannel;
+}
